requirements.txt
# HawkAI - Dependências
# Instalar com: pip install -r requirements.txt
# Sistema 100% local usando Ollama

# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# Local LLM (Ollama) - 100% local, sem API keys
langchain-ollama>=0.2.0

# Vector Store
chromadb>=0.5.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
streamlit>=1.38.0

# Data Processing
pandas>=2.0.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.0.0
